Move Select HOCs in a separate folder
