+++ packages/react-vapor/src/components/filterBox/tests/FilterBoxSelectors.spec.ts
@@ -1,6 +1,6 @@
 import {IReactVaporState} from '../../../ReactVapor';
 import {defaultMatchFilter, MatchFilter} from '../../../utils/FilterUtils';
-import {ISelectWithFilterProps} from '../../select/SelectWithFilter';
+import {ISelectWithFilterProps} from '../../select/hoc/withFilter';
 import {IFilterState} from '../FilterBoxReducers';
 import {FilterBoxSelectors} from '../FilterBoxSelectors';
