+++ packages/react-vapor/src/components/select/SelectSelector.ts
@@ -8,9 +8,9 @@ import {DropSelectors} from '../drop/redux/DropReducers';
 import {FilterBoxSelectors} from '../filterBox/FilterBoxSelectors';
 import {IItemBoxProps} from '../itemBox/ItemBox';
 import {IListBoxState} from '../listBox/ListBoxReducers';
+import {ISelectWithFilterProps} from './hoc/withFilter';
 import {IMultiSelectProps} from './MultiSelectConnected';
 import {ISelectProps, SelectConnected} from './SelectConnected';
-import {ISelectWithFilterProps} from './SelectWithFilter';
 
 const getListState = (state: IReactVaporState, ownProps: ISelectProps): string[] =>
     state.selectWithFilter && state.selectWithFilter[ownProps.id] ? state.selectWithFilter[ownProps.id].list : [];
